requirements.txt
autogen~=0.2.0
google-cloud-sql~=1.0.0
google-cloud-storage~=2.0.0
mysql-connector-python~=8.0.0
pandas~=2.0.0
numpy~=1.26.0
sqlalchemy~=2.0.0
python-dotenv~=1.0.0